+++ MoltenVK/MoltenVK/Utility/MVKFoundation.h
@@ -379,9 +379,6 @@ static constexpr bool mvkVkComponentMappingsMatch(VkComponentMapping cm1, VkComp
 			mvkVKComponentSwizzlesMatch(cm1.a, cm2.a, VK_COMPONENT_SWIZZLE_A));
 }
 
-/** Print the size of the type. */
-#define mvkPrintSizeOf(type)    printf("Size of " #type " is %lu.\n", sizeof(type))
-
 
 #pragma mark Math